fix textInputLayout bug & fragment page jump bug

+++ app/src/main/java/com/example/receiptpocket/pocket/qrscan/manualInput/Interactors/ManualInputInteractorImpl.kt
@@ -14,20 +14,45 @@ class ManualInputInteractorImpl: ManualInputInteractor {
     ) {
         Thread{
 
-            try {
-                val receiptDao = ReceiptDatabase.getDatabase(App.appContext).receiptDao()
+            // 確認輸入欄位資料符合規定
+            if(isCode1TwoChar(item.code_1) && isCode1Uppercase(item.code_1) && isCode2EightChar(item.code_2)){
+                try {
+                    val receiptDao = ReceiptDatabase.getDatabase(App.appContext).receiptDao()
 
-                if (isDelete) { receiptDao.delete(item) }
-                else { receiptDao.insert(item) }
+                    if (isDelete) { receiptDao.delete(item) }
+                    else { receiptDao.insert(item) }
 
-                listener.onUpdateDataSuccess()
+                    listener.onUpdateDataSuccess()
 
-            } catch (e: SQLException){
-                e.printStackTrace()
+                } catch (e: SQLException){
+                    e.printStackTrace()
+                }
+            }
+            else{
+
+                if (!isCode1TwoChar(item.code_1)) { listener.setCode1CharNumError() }
+                else if (!isCode1Uppercase(item.code_1)) { listener.setCode1UppercaseError() }
+                else { listener.setCode2CharNumError() }
             }
 
         }.start()
 
+    }
+
+
+    private fun isCode1TwoChar(str: String): Boolean {
+        if(str.length==2) return true
+        return false
+    }
+
+    private fun isCode2EightChar(str: String): Boolean {
+        if(str.length==8) return true
+        return false
+    }
+
+    private fun isCode1Uppercase(str: String): Boolean {
+        if(str[0] in 'A'..'Z' && str[1] in 'A'..'Z') return true
+        return false
     }
 
 }
